Implemented JWT Login
Brought in @auth0/angular-jwt library and set to grab and include jwt on all requests.
Implemented interceptor to grab login and validate or return error.
Set getToken method to logout if token is expired
Token is currently just a hardcoded string since there arent any good libraries to generate it in angular and it will be handled by backend anyway